+++ clients/admin-ui/src/features/common/api.slice.ts
@@ -51,6 +51,7 @@ export const baseApi = createApi({
     "Manual Fields",
     "Manual Tasks",
     "Messaging Templates",
+    "Messaging Config",
     "Dictionary",
     "System Vendors",
     "Latest Scan",